requirements.txt
pdf2image>=1.16.3
Pillow>=10.0.0
openai>=1.0.0
pydantic>=2.0.0
tqdm>=4.66.0
PyYAML>=6.0